# backend/api/controllers/disputeController.js
/**
 * Dispute Controller
 *
 * Handles dispute resolution, evidence management, and appeals.
 * Evidence files are stored on IPFS with virus scanning and thumbnail generation.
 */

import prisma from '../../lib/prisma.js';
import { buildPaginatedResponse, parsePagination } from '../../lib/pagination.js';
import { uploadEvidence } from '../middleware/fileUpload.js';
import ipfsService from '../../services/ipfsService.js';
import { broadcastToDispute } from '../websocket/handlers.js';
import { verifyFile, merkleRoot, hashFile } from '../../services/ipfsHashService.js';
import { submitTransaction } from '../../services/stellarService.js';

/**
 * List and get handlers assume query/params are already validated by
 * `validate(disputeListQueryRules)` and `validate(disputeEscrowIdParamRules)`.
 */

const VALID_DISPUTE_SORT_FIELDS = new Set(['raisedAt', 'resolvedAt', 'id']);
const VALID_SORT_ORDERS = new Set(['asc', 'desc']);

const DISPUTE_MAX_LIMIT = 50;

const listDisputes = async (req, res) => {
  try {
    const { page, skip } = parsePagination(req.query);
    const limit = Math.min(parsePagination(req.query).limit, DISPUTE_MAX_LIMIT);
    const {
      status,
      raisedBy,
      dateFrom,
      dateTo,
      sortBy = 'raisedAt',
      sortOrder = 'desc',
    } = req.query;

    if (!VALID_DISPUTE_SORT_FIELDS.has(sortBy)) {
      return res.status(400).json({
        error: 'Invalid sortBy value',
        allowed: [...VALID_DISPUTE_SORT_FIELDS],
      });
    }
    if (!VALID_SORT_ORDERS.has(sortOrder)) {
      return res.status(400).json({
        error: 'Invalid sortOrder value',
        allowed: [...VALID_SORT_ORDERS],
      });
    }

    if (dateFrom && isNaN(Date.parse(dateFrom))) {
      return res.status(400).json({ error: 'dateFrom must be a valid ISO date string' });
    }
    if (dateTo && isNaN(Date.parse(dateTo))) {
      return res.status(400).json({ error: 'dateTo must be a valid ISO date string' });
    }
    if (dateFrom && dateTo && new Date(dateFrom) > new Date(dateTo)) {
      return res.status(400).json({ error: 'dateFrom must not be after dateTo' });
    }

    const where = {
      tenantId: req.tenant.id,
    };

    if (status === 'resolved') {
      where.resolvedAt = { not: null };
    } else if (status === 'unresolved') {
      where.resolvedAt = null;
    }

    if (raisedBy) where.raisedByAddress = raisedBy;
    if (dateFrom || dateTo) {
      where.raisedAt = {};
      if (dateFrom) where.raisedAt.gte = new Date(dateFrom);
      if (dateTo) where.raisedAt.lte = new Date(dateTo);
    }

    const [disputes, total] = await Promise.all([
      prisma.dispute.findMany({
        where,
        include: {
          escrow: {
            select: {
              clientAddress: true,
              freelancerAddress: true,
              totalAmount: true,
              status: true,
            },
          },
          evidence: {
            select: {
              id: true,
              evidenceType: true,
              submittedBy: true,
              submittedAt: true,
              filename: true,
              ipfsCid: true,
              thumbnailCid: true,
              scanStatus: true,
            },
            orderBy: { submittedAt: 'desc' },
          },
          _count: {
            select: { evidence: true, appeals: true },
          },
        },
        orderBy: { [sortBy]: sortOrder },
        skip,
        take: limit,
      }),
      prisma.dispute.count({ where }),
    ]);

    const response = buildPaginatedResponse(disputes, {
      total,
      page,
      limit,
    });

    res.json(response);
  } catch (error) {
    console.error('Error listing disputes:', error);
    res.status(500).json({ error: 'Failed to list disputes' });
  }
};

const getDispute = async (req, res) => {
  try {
    const { escrowId } = req.params;
    const escrowIdBigInt = BigInt(escrowId);

    const dispute = await prisma.dispute.findFirst({
      where: {
        escrowId: escrowIdBigInt,
        tenantId: req.tenant.id,
      },
      include: {
        escrow: {
          select: {
            clientAddress: true,
            freelancerAddress: true,
            arbiterAddress: true,
            tokenAddress: true,
            totalAmount: true,
            remainingBalance: true,
            status: true,
            deadline: true,
            createdAt: true,
          },
        },
        evidence: {
          include: {
            _count: true,
          },
          orderBy: { submittedAt: 'desc' },
        },
        appeals: {
          orderBy: { createdAt: 'desc' },
        },
      },
    });

    if (!dispute) {
      return res.status(404).json({ error: 'Dispute not found' });
    }

    const evidenceWithUrls = await Promise.all(
      dispute.evidence.map(async (evidence) => {
        const evidenceData = { ...evidence };
        if (evidence.ipfsCid) {
          evidenceData.fileUrl = await ipfsService.getFileUrl(evidence.ipfsCid);
        }
        if (evidence.thumbnailCid) {
          evidenceData.thumbnailUrl = await ipfsService.getFileUrl(evidence.thumbnailCid);
        }
        return evidenceData;
      }),
    );

    res.json({
      ...dispute,
      evidence: evidenceWithUrls,
    });
  } catch (error) {
    console.error('Error getting dispute:', error);
    res.status(500).json({ error: 'Failed to get dispute' });
  }
};

const postEvidence = async (req, res) => {
  try {
    const { description, role } = req.body;
    const userAddress = req.userAddress; // set by validateDisputeAccess in fileUpload middleware

    if (!userAddress) {
      return res.status(401).json({ error: 'User not authenticated' });
    }

    const dispute = req.dispute;
    const uploadResults = req.ipfsUploadResults || [];
    const scanResults = req.virusScanResults || [];

    if (uploadResults.length === 0 && !description) {
      return res.status(400).json({
        error: 'No evidence provided',
        message: 'Either files or text description is required',
      });
    }

    const evidenceRecords = [];

    for (const uploadResult of uploadResults) {
      const scanResult = scanResults.find((s) => s.originalname === uploadResult.originalname);

      const evidenceRecord = await prisma.disputeEvidence.create({
        data: {
          tenantId: req.tenant.id,
          disputeId: dispute.id,
          submittedBy: userAddress,
          role: role || determineUserRole(dispute, userAddress),
          evidenceType: uploadResult.metadata.mimeType.startsWith('image/') ? 'image' : 'file',
          content: uploadResult.ipfsCid,
          description: description || null,
          filename: uploadResult.originalname,
          mimeType: uploadResult.mimetype,
          fileSize: uploadResult.size,
          ipfsCid: uploadResult.ipfsCid,
          thumbnailCid: uploadResult.thumbnailCid,
          scanStatus: scanResult?.status || 'pending',
          scanResult: JSON.stringify(scanResult) || null,
        },
      });

      evidenceRecords.push(evidenceRecord);
    }

    if (description && uploadResults.length === 0) {
      const textEvidence = await prisma.disputeEvidence.create({
        data: {
          tenantId: req.tenant.id,
          disputeId: dispute.id,
          submittedBy: userAddress,
          role: role || determineUserRole(dispute, userAddress),
          evidenceType: 'text',
          content: description,
          description: null,
        },
      });

      evidenceRecords.push(textEvidence);
    }

    const evidenceWithUrls = await Promise.all(
      evidenceRecords.map(async (evidence) => {
        const evidenceData = { ...evidence };
        if (evidence.ipfsCid) {
          evidenceData.fileUrl = await ipfsService.getFileUrl(evidence.ipfsCid);
        }
        if (evidence.thumbnailCid) {
          evidenceData.thumbnailUrl = await ipfsService.getFileUrl(evidence.thumbnailCid);
        }
        return evidenceData;
      }),
    );

    broadcastToDispute(dispute.id, {
      type: 'evidence_added',
      disputeId: dispute.id,
      evidence: evidenceWithUrls,
      submittedBy: userAddress,
      timestamp: new Date().toISOString(),
    });

    res.status(201).json({
      message: 'Evidence uploaded successfully',
      evidence: evidenceWithUrls,
      count: evidenceRecords.length,
    });
  } catch (error) {
    console.error('Error posting evidence:', error);
    res.status(500).json({ error: 'Failed to post evidence' });
  }
};

const listEvidence = async (req, res) => {
  try {
    const { id } = req.params;
    const { page, limit, skip } = parsePagination(req.query);
    const { evidenceType, submittedBy } = req.query;

    const where = {
      tenantId: req.tenant.id,
      disputeId: parseInt(id),
    };

    if (evidenceType) where.evidenceType = evidenceType;
    if (submittedBy) where.submittedBy = submittedBy;

    const [evidence, total] = await Promise.all([
      prisma.disputeEvidence.findMany({
        where,
        orderBy: { submittedAt: 'desc' },
        skip,
        take: limit,
      }),
      prisma.disputeEvidence.count({ where }),
    ]);

    const evidenceWithUrls = await Promise.all(
      evidence.map(async (evidenceItem) => {
        const evidenceData = { ...evidenceItem };
        if (evidenceItem.ipfsCid) {
          evidenceData.fileUrl = await ipfsService.getFileUrl(evidenceItem.ipfsCid);
        }
        if (evidenceItem.thumbnailCid) {
          evidenceData.thumbnailUrl = await ipfsService.getFileUrl(evidenceItem.thumbnailCid);
        }
        return evidenceData;
      }),
    );

    const response = buildPaginatedResponse(evidenceWithUrls, {
      total,
      page,
      limit,
    });

    res.json(response);
  } catch (error) {
    console.error('Error listing evidence:', error);
    res.status(500).json({ error: 'Failed to list evidence' });
  }
};

const autoResolve = async (req, res) => {
  try {
    const dispute = req.dispute;

    // Run the rules engine to compute the resolution split
    const { runAutomatedResolution } = await import('../../services/disputeResolution.js');
    const resolutionResult = await runAutomatedResolution(dispute.id);

    // Broadcast the ruling on-chain if the relayer is configured
    if (process.env.RELAYER_SECRET_KEY && process.env.ESCROW_CONTRACT_ID) {
      try {
        const { createRelayer } = await import('../../services/relayerService.js');
        const relayer = createRelayer({
          network:
            process.env.STELLAR_NETWORK === 'mainnet'
              ? 'Public Global Stellar Network ; September 2015'
              : 'Test SDF Network ; September 2015',
          contractId: process.env.ESCROW_CONTRACT_ID,
          relayerSecret: process.env.RELAYER_SECRET_KEY,
        });

        const updatedDispute = resolutionResult.dispute;
        await relayer.executeMetaTransaction({
          function: 'resolve_dispute',
          args: {
            escrow_id: dispute.escrowId.toString(),
            client_amount: updatedDispute.clientAmount ?? '0',
            freelancer_amount: updatedDispute.freelancerAmount ?? '0',
          },
        });
      } catch (onChainErr) {
        // Log but do not fail the response — the DB resolution is committed;
        // the indexer will reconcile on-chain state on its next poll.
        console.error('autoResolve: on-chain broadcast failed', onChainErr.message);
      }
    }

    broadcastToDispute(dispute.id, {
      type: 'dispute_resolved',
      disputeId: dispute.id,
      resolution: resolutionResult.dispute,
      timestamp: new Date().toISOString(),
    });

    res.json({
      message: 'Dispute auto-resolved successfully',
      resolution: resolutionResult.dispute,
      resolutionType: resolutionResult.resolutionType,
    });
  } catch (error) {
    console.error('Error auto-resolving dispute:', error);
    res.status(500).json({ error: 'Failed to auto-resolve dispute' });
  }
};

const getRecommendation = async (req, res) => {
  try {
    const dispute = req.dispute;

    const evidence = await prisma.disputeEvidence.findMany({
      where: {
        disputeId: dispute.id,
        tenantId: req.tenant.id,
      },
      orderBy: { submittedAt: 'desc' },
    });

    const recommendation = generateResolutionRecommendation(dispute, evidence);

    res.json({
      disputeId: dispute.id,
      recommendation,
      evidenceCount: evidence.length,
      generatedAt: new Date().toISOString(),
    });
  } catch (error) {
    console.error('Error getting recommendation:', error);
    res.status(500).json({ error: 'Failed to get recommendation' });
  }
};

const postAppeal = async (req, res) => {
  try {
    const { reason } = req.body;
    const userAddress = req.user?.walletAddress ?? req.userAddress;

    if (!reason || reason.trim().length === 0) {
      return res.status(400).json({ error: 'Appeal reason is required' });
    }

    const dispute = req.dispute;

    const appeal = await prisma.disputeAppeal.create({
      data: {
        tenantId: req.tenant.id,
        disputeId: dispute.id,
        appealedBy: userAddress,
        reason: reason.trim(),
      },
    });

    broadcastToDispute(dispute.id, {
      type: 'appeal_filed',
      disputeId: dispute.id,
      appealId: appeal.id,
      appealedBy: userAddress,
      timestamp: new Date().toISOString(),
    });

    res.status(201).json({
      message: 'Appeal filed successfully',
      appeal,
    });
  } catch (error) {
    console.error('Error posting appeal:', error);
    res.status(500).json({ error: 'Failed to post appeal' });
  }
};

const patchAppeal = async (req, res) => {
  try {
    const { appealId } = req.params;
    const { status, reviewNotes } = req.body;
    const userAddress = req.user?.walletAddress ?? req.userAddress;

    const appeal = await prisma.disputeAppeal.findFirst({
      where: {
        id: parseInt(appealId),
        tenantId: req.tenant.id,
      },
    });

    if (!appeal) {
      return res.status(404).json({ error: 'Appeal not found' });
    }

    const updatedAppeal = await prisma.disputeAppeal.update({
      where: { id: appeal.id },
      data: {
        status,
        reviewNotes,
        reviewedBy: userAddress,
        resolvedAt: status === 'approved' || status === 'rejected' ? new Date() : null,
      },
    });

    res.json({
      message: 'Appeal updated successfully',
      appeal: updatedAppeal,
    });
  } catch (error) {
    console.error('Error updating appeal:', error);
    res.status(500).json({ error: 'Failed to update appeal' });
  }
};

const getResolutionHistory = async (req, res) => {
  try {
    const { page, limit, skip } = parsePagination(req.query);

    const [disputes, total] = await Promise.all([
      prisma.dispute.findMany({
        where: {
          tenantId: req.tenant.id,
          resolvedAt: { not: null },
        },
        include: {
          escrow: {
            select: {
              clientAddress: true,
              freelancerAddress: true,
              totalAmount: true,
            },
          },
        },
        orderBy: { resolvedAt: 'desc' },
        skip,
        take: limit,
      }),
      prisma.dispute.count({
        where: {
          tenantId: req.tenant.id,
          resolvedAt: { not: null },
        },
      }),
    ]);

    const response = buildPaginatedResponse({
      items: disputes,
      total,
      page,
      limit,
      request: req,
    });

    res.json(response);
  } catch (error) {
    console.error('Error getting resolution history:', error);
    res.status(500).json({ error: 'Failed to get resolution history' });
  }
};

/**
 * POST /api/disputes/verify-evidence
 * Body: multipart/form-data with files[] + evidenceIds[]
 *
 * Re-uploads files and verifies their SHA-256 hashes against stored values.
 * Also recomputes the Merkle root and compares against the stored root.
 */
const verifyEvidence = async (req, res) => {
  try {
    const { evidenceIds } = req.body;
    const files = req.files ?? [];

    if (!evidenceIds || !files.length) {
      return res.status(400).json({ error: 'evidenceIds and files are required' });
    }

    const ids = (Array.isArray(evidenceIds) ? evidenceIds : [evidenceIds]).map(Number);

    const records = await prisma.disputeEvidence.findMany({
      where: { id: { in: ids } },
      select: { id: true, fileHash: true, merkleRoot: true, filename: true },
    });

    if (records.length !== files.length) {
      return res.status(400).json({ error: 'File count must match evidenceIds count' });
    }

    const fileResults = files.map((file, i) => {
      const record = records[i];
      if (!record?.fileHash) return { id: record?.id, valid: false, reason: 'No stored hash' };
      const { hex } = hashFile(file.buffer);
      return { id: record.id, filename: file.originalname, valid: hex === record.fileHash };
    });

    const storedHashes = records.map((r) => r.fileHash).filter(Boolean);
    const recomputedRoot = merkleRoot(storedHashes);
    const storedRoot = records[0]?.merkleRoot ?? null;
    const rootMatch = storedRoot ? recomputedRoot === storedRoot : null;

    return res.json({
      allValid: fileResults.every((r) => r.valid),
      fileResults,
      merkleRoot: recomputedRoot,
      storedMerkleRoot: storedRoot,
      rootMatch,
    });
  } catch (err) {
    return res.status(500).json({ error: err.message });
  }
};

function determineUserRole(dispute, userAddress) {
  if (dispute.raisedByAddress === userAddress) {
    return dispute.escrow.clientAddress === userAddress ? 'client' : 'freelancer';
  }
  if (dispute.escrow.clientAddress === userAddress) return 'client';
  if (dispute.escrow.freelancerAddress === userAddress) return 'freelancer';
  return 'arbiter';
}

function generateResolutionRecommendation(dispute, evidence) {
  const clientEvidence = evidence.filter((e) => e.role === 'client').length;
  const freelancerEvidence = evidence.filter((e) => e.role === 'freelancer').length;
  const fileEvidence = evidence.filter(
    (e) => e.evidenceType === 'file' || e.evidenceType === 'image',
  ).length;

  let recommendation = {
    suggestedOutcome: 'manual_review',
    confidence: 0.5,
    reasoning: [],
  };

  if (fileEvidence > 0) {
    recommendation.confidence += 0.2;
    recommendation.reasoning.push('Documentary evidence provided');
  }

  if (clientEvidence > freelancerEvidence + 2) {
    recommendation.suggestedOutcome = 'favor_client';
    recommendation.confidence += 0.1;
    recommendation.reasoning.push('Client provided significantly more evidence');
  } else if (freelancerEvidence > clientEvidence + 2) {
    recommendation.suggestedOutcome = 'favor_freelancer';
    recommendation.confidence += 0.1;
    recommendation.reasoning.push('Freelancer provided significantly more evidence');
  }

  recommendation.confidence = Math.min(recommendation.confidence, 0.9);

  return recommendation;
}

export default {
  listDisputes,
  getDispute,
  postEvidence,
  listEvidence,
  autoResolve,
  getRecommendation,
  postAppeal,
  patchAppeal,
  getResolutionHistory,
  uploadEvidence,
  verifyEvidence,
};