# backend/openapi.yaml
openapi: 3.0.3
info:
  title: TrustHood Escrow API
  version: 0.1.0
  description: |
    REST API for the TrustHood Escrow platform — a trust-minimised escrow
    service built on the Stellar blockchain.

    ## Authentication
    Most endpoints require a Bearer JWT token in the `Authorization` header.
    Admin endpoints additionally require the `x-admin-api-key` header.

    ## Pagination
    All collection endpoints return a standardised envelope:
    ```
    {
      "data": [...],
      "page": 1, "limit": 20, "total": 42,
      "totalPages": 3, "hasNextPage": true, "hasPreviousPage": false
    }
    ```

    ## Environments
    - **Testnet**: `https://api.test.trusthood-escrow.example.com`
    - **Mainnet**: `https://api.trusthood-escrow.example.com`
  contact:
    name: TrustHood Escrow
    url: https://trusthood-escrow.example.com
  license:
    name: AGPL-3.0
    url: https://www.gnu.org/licenses/agpl-3.0.en.html
  x-repository: https://github.com/trusthood-escrow/trusthood-escrow
  x-docs-url: https://docs.trusthood-escrow.example.com
servers:
  - url: http://localhost:4000
    description: Local development
  - url: https://api.test.trusthood-escrow.example.com
    description: Testnet
  - url: https://api.trusthood-escrow.example.com
    description: Mainnet
...
security:
  - BearerAuth: []
paths:
...
  /api/disputes:
    get:
      tags: [Disputes]
      summary: List disputes with pagination
      operationId: listDisputes
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/Limit'
      responses:
        '200':
          description: Paginated dispute list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedDisputes'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /api/disputes/{escrowId}:
    get:
      tags: [Disputes]
      summary: Get dispute details for a specific escrow
      operationId: getDispute
      security:
        - BearerAuth: []
      parameters:
        - name: escrowId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Dispute detail
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Dispute'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT access token obtained from /api/auth/login
    AdminAuth:
      type: apiKey
      in: header
      name: x-admin-api-key
      description: Admin API key for protected admin endpoints
  parameters:
    Page:
      name: page
      in: query
      description: 1-based page number. Values below 1 are normalised to 1.
      schema:
        type: integer
        default: 1
        minimum: 1
    Limit:
      name: limit
      in: query
      description: Page size. Defaults to 20 and is capped at 100.
      schema:
        type: integer
        default: 20
        minimum: 1
        maximum: 100
...
  schemas:
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
          example: Something went wrong
        errorId:
          type: string
          description: Sentry event ID for support correlation
    PaginatedResponse:
      type: object
      required: [data, page, limit, total, totalPages, hasNextPage, hasPreviousPage]
      properties:
        data:
          type: array
          items: {}
        page:
          type: integer
          example: 1
        limit:
          type: integer
          example: 20
        total:
          type: integer
          example: 42
        totalPages:
          type: integer
          example: 3
        hasNextPage:
          type: boolean
          example: true
        hasPreviousPage:
          type: boolean
          example: false
...
    Dispute:
      type: object
      properties:
        id:
          type: string
        escrowId:
          type: string
        raisedBy:
          type: string
        reason:
          type: string
        status:
          type: string
          enum: [Open, Resolved]
        resolution:
          type: object
          nullable: true
          properties:
            clientAmount:
              type: string
            freelancerAmount:
              type: string
            notes:
              type: string
        createdAt:
          type: string
          format: date-time
        resolvedAt:
          type: string
          format: date-time
          nullable: true
    PaginatedDisputes:
      allOf:
        - $ref: '#/components/schemas/PaginatedResponse'
        - type: object
          properties:
            data:
              type: array
              items:
                $ref: '#/components/schemas/Dispute'
...
  responses:
    Unauthorized:
      description: Missing or invalid authentication token
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...